- fixed comments after review
